+++ src/commands/diff.ts
@@ -1,8 +1,54 @@
 import * as vscode from 'vscode';
 import * as parsers from './../parsers';
 import ForceCodeContentProvider from '../providers/ContentProvider';
+import { ForcecodeCommand } from './forcecodeCommand';
+import { commandService } from '../services';
 
 const PROVIDER: string = 'forcecode://salesforce.com';
+
+export class DiffContext extends ForcecodeCommand {
+  constructor() {
+    super();
+    this.commandName = 'ForceCode.diff';
+    this.hidden = true;
+  }
+
+  public command(context, selectedResource?) {
+    return commandService.runCommand('ForceCode.diffMenu', context, selectedResource);
+  }
+}
+
+export class DiffMenu extends ForcecodeCommand {
+  constructor() {
+    super();
+    this.commandName = 'ForceCode.diffMenu';
+    this.name = 'Diffing file'; //+ getFileName(vscode.window.activeTextEditor.document),
+    this.hidden = false;
+    this.description = 'Diff the current file with what is on the server';
+    this.detail = 'Diff the file';
+    this.icon = 'diff';
+    this.label = 'Diff';
+  }
+
+  public command(context, selectedResource?) {
+    if (selectedResource && selectedResource.path) {
+      return vscode.workspace.openTextDocument(selectedResource).then(doc => diff(doc));
+    }
+    if (!vscode.window.activeTextEditor) {
+      return;
+    }
+    const ttype: string | undefined = parsers.getToolingType(
+      vscode.window.activeTextEditor.document
+    );
+    if (!ttype) {
+      throw { message: 'Metadata type not supported for diffing' };
+    }
+    return diff(
+      vscode.window.activeTextEditor.document,
+      ttype === 'AuraDefinition' || ttype === 'LightningComponentResource'
+    );
+  }
+}
 
 export default function diff(document: vscode.TextDocument, auraSource?: boolean) {
   if (!document) {
